Debugging ADC interface, more features, ...[WIP]
Apparently the first sample of a pair ships on the
negative clock edge, the second one on the following
positive edge. Depending whether we send muxed data
on port B or port A the B or A sample is shipped
first, respectively, but always the first sample of
a pair is latched on the negative edge.